trade_lab/requirements.txt
streamlit>=1.41.0
pandas>=2.2.0
numpy>=1.26.0
PyYAML>=6.0.0
